Count erroring tests separately from failed ones
The junit reporter separates failed and error test results by looking at
the spec description, which is a string for failed specs and a lisp sexp
for errors.

For specs that are recognized as errors, the reporter sets the spec's
status slot to `error'.  A error counting function
`buttercup-suites-total-specs-error' had to be added to count the errors
and separate them from failures.